+++ nf_core/pipelines/lint/schema_description.py
@@ -1,4 +1,4 @@
-import nf_core.schema
+import nf_core.pipelines.schema
 
 
 def schema_description(self):
@@ -17,7 +17,7 @@ def schema_description(self):
 
     # First, get the top-level config options for the pipeline
     # Schema object already created in the `schema_lint` test
-    self.schema_obj = nf_core.schema.PipelineSchema()
+    self.schema_obj = nf_core.pipelines.schema.PipelineSchema()
     self.schema_obj.get_schema_path(self.wf_path)
     self.schema_obj.get_wf_params()
     self.schema_obj.no_prompts = True